# specs/api.yaml
openapi: 3.0.0
info:
  title: luna-marketplace
  version: 1.0.0
servers:
  - url: https://{{baseurl}}
...
paths:
...
  /admin:
    post:
      tags:
        - admin
      summary: Signup Admin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpAdminRequest'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignUpAdminResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                type: object
              example:
                error: true
                code: 403
                message: Public address not accepted to be an admin
    get:
      tags:
        - admin
      summary: Get Admin
      operationId: GetAdmins
      parameters:
        - name: public_address
          in: query
          schema:
            type: string
          required: false
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAdminsResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
              example:
                error: true
                code: 404
                message: Token metadata was not found
...
components:
  schemas:
    GenericResponse:
      properties:
        error:
          type: boolean
          example: false
        code:
          type: number
          example: 200
        message:
          type: string
          example: Successful
...
    SignUpAdminRequest:
      type: object
      properties:
        public_address:
          type: string
    SignUpAdminResponse:
      allOf:
        - $ref: '#/components/schemas/GenericResponse'
        - type: object
          properties:
            public_address:
              type: string
              example: publicAddress
    AdminData:
      type: object
      properties:
        public_address:
          type: string
        username:
          type: string
        nonce:
          type: number
      example:
        public_address: publicAddress
        username: lekky
        nonce: 345056
    GetAdminsResponse:
      allOf:
        - $ref: '#/components/schemas/GenericResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/AdminData'
...